feat(payment): accept correct change only
Update the hopper calculation to complete the purchase when exact
payment+change can be made.

The hopper calculates how to change for the product and give change,
using the tray and hopper coins. This could mean giving change from the
unused coins in the tray, as well as from the hopper.